Move SmellRepository to top level.
